Time: 16 ms (78.46%), Space: 21.7 MB (30.09%) - LeetHub

+++ 0098-validate-binary-search-tree/0098-validate-binary-search-tree.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool res = 1;
-    void dfs(TreeNode* root, int64_t left, int64_t right){
+    void dfs(TreeNode* root, long long int left, long long int right){
         if(!root){
             return;
         }
